FEAT: add xorbits experimental
