+++ doc/source/reference/experimental/dedup.rst
@@ -1,0 +1,12 @@
+.. _api.dedup:
+
+=========
+Dedup API
+=========
+.. currentmodule:: xorbits.experimental
+
+
+.. autosummary::
+   :toctree: generated/
+
+   dedup
