+++ doc/source/reference/experimental/index.rst
@@ -1,0 +1,10 @@
+.. _experimental_api:
+
+================
+Experimental API
+================
+
+.. toctree::
+   :maxdepth: 2
+
+   dedup
